test(api): organize pytest suite and add test scaffolding
