+++ apps/api/tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from django.test import RequestFactory
+
+
+@pytest.fixture
+def request_factory() -> RequestFactory:
+    return RequestFactory()
