+++ apps/api/tests/factories.py
@@ -1,0 +1,5 @@
+"""
+Shared test factory scaffolding for backend tests.
+
+Concrete model factories will be added as domain models are introduced.
+"""
